Separate evaluator, remove cranelift (#5)
This takes us a few steps closer to adding support for user-defined functions and async. In this PR, the evaluator moves into its own file and can accept multiple codegen'd functions.

This allows us to build up a single vector of the instructions of the user's script and run these instructions without recursion, which will come in handy once we add user-defined functions.

It will also come in handy as we'll have a single entry-point for async evaluation that should be easy to manage.