Batch of parser simplifications (#10)
A batch of parser simplifications:

1) I noticed that the way I was doing lazy lexing was actually pretty
bad in practice and a single token may be re-lexed dozens of times.
Oops. Moved the lexer to being eager instead of lazy so we just lex once
and then work over the lexed tokens.

2) Simplified how the lexer tokens were handled in the parser, making
them Copy in the process

3) Since we're not aiming for a REPL this time around, go ahead and
remove some of the REPL-like logic with the delta. Now just have a batch
of parser results that things can use after the parser finishes.

4) Remove some lifetimes I'd thrown in there to be more efficient, but
they end up making refactoring harder. They should be an optimisation
step anyway, so their use was a bit premature.

5) Cleanup the Typechecker setup in the main loop so we don't have to
create it twice.