fix segfault when returning custom types from scripts (#81)
Prior to this PR, we would run into double frees whenever returning a
custom type from a script. The reason for this was that the drop
implementation for the engine would check every register to see if it
had a value that needed to be dropped, but there was a bug in the
implementation. The drop would check _all_ registers, including the
return register, causing it to free the value it was returning, which
was freed for a second time when the user dropped it.

This PR resolves the issue by skipping the `0` register in the
evaluator's drop implementation.